added book manager class
